fastapi/pyproject.toml
[tool.poetry]
name = "modular-rag"
version = "0.1.0"
description = "A modular RAG system built with FastAPI and LangChain"
authors = ["syshin0116 <syshin0116@gmail.com>"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/syshin0116/modular-rag"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "*"
langchain = "*"
uvicorn = "*"
pydantic = {extras = ["email"], version = "^2.9.2"}
requests = "*"
redis = "*"
pydantic-settings = "^2.6.0"
python-jose = "^3.3.0"
aiomysql = "^0.2.0"
itsdangerous = "^2.2.0"
jinja2 = "^3.1.4"

[tool.poetry.dev-dependencies]
pytest = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"